# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecommerce APIs Documentation",
        "description": "Ecommerce mobile and Web Frontend APIs",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Ecommerce API Server"
        }
    ],
    "paths": {
...
        "/api/product": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get List of Products",
                "description": "Return List of Products",
                "operationId": "getAllProducts",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Store product in DB",
                "description": "Store product in DB",
                "operationId": "storeProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "category_id",
                                    "vendor_id",
                                    "price",
                                    "quantity"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Jacket"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": 1
                                    },
                                    "vendor_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": 1
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "integer",
                                        "example": 10.23
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": 100
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Jacket description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Unprocessable content"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/product/{product}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Show product",
                "description": "Show product",
                "operationId": "getProduct",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Id of product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Unprocessable content"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update product",
                "description": "Update product",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Id of product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Jacket"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": 1
                                    },
                                    "vendor_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": 1
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "integer",
                                        "example": 10.23
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": 100
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Jacket description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Unprocessable content"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete Products",
                "description": "Delete Products",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Id of Product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    }
}